kiratech_inventory_management/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.7
Django==4.1.2
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==5.0.0
Markdown==3.4.1
mysqlclient==2.1.1
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
zipp==3.9.0
